[642] Changes to separate template for download/upload and read-only download.
Changes the edit page to show two separate tools.  The download tool is now a
read-only view of the current datasets and inventory items on DGUK.

The upload tool now provides a template that users should download (and a link to the guidance doc - not yet there) and then re-upload when modified.